// src/typings/types.ts
export type CharacterJson = {
  strokes: string[];
  medians: number[][][];
  radStrokes?: number[];
};